backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.65.1
FastAPI-SQLAlchemy==0.2.1
frozenlist==1.3.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
pip==21.3.1
psycopg2==2.9.3
pydantic==1.9.2
python-dotenv==0.20.0
setuptools==60.2.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.14.2
typing_extensions==4.3.0
uvicorn==0.14.0
wheel==0.37.1
yarl==1.8.1
zipp==3.8.1